legacy tests now run using BaseTest
